+++ packages/ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@teselagen/ui",
-  "version": "0.5.23-beta.16",
+  "version": "0.5.23-beta.17",
   "main": "./src/index.js",
   "type": "module",
   "exports": {
